+++ changelog.md
@@ -3,6 +3,8 @@
 - `rio.Dropdown` will now open a fullscreen popup on mobile devices
 - `riol.MediaPlayer` now also triggers the `on_playback_end` event when the
     video loops
+- experimental support for base-URL
+- dialogs can now store a result value similar to futures
 
 ## 0.9.2